+++ src/test/java/org/folio/rest/api/InstanceStorageTest.java
@@ -662,6 +662,63 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
   }
 
   @Test
+  public void canRetrieveAllInstances() throws InterruptedException, ExecutionException, TimeoutException {
+    var firstInstanceId = UUID.randomUUID();
+    var firstInstanceToCreate = smallAngryPlanet(firstInstanceId);
+    var secondInstanceId = UUID.randomUUID();
+    var secondInstanceToCreate = nod(secondInstanceId);
+
+    createInstance(firstInstanceToCreate);
+    createInstance(secondInstanceToCreate);
+
+    var query = "(cql.allRecords=1) sortBy title";
+    var retrieveCompleted = new CompletableFuture<Response>();
+    var retrieveByTitleCompleted = new CompletableFuture<Response>();
+
+    getClient().post(instancesStorageUrl("/retrieve"), new JsonObject(), TENANT_ID,
+      json(retrieveCompleted));
+    getClient().post(instancesStorageUrl("/retrieve"), new JsonObject().put("query", query), TENANT_ID,
+      json(retrieveByTitleCompleted));
+
+    var retrieveBody = retrieveCompleted.get(10, SECONDS).getJson();
+    var allInstances = retrieveBody.getJsonArray(INSTANCES_KEY);
+
+    var retrieveByTitleBody = retrieveByTitleCompleted.get(10, SECONDS).getJson();
+    var sortedInstances = retrieveByTitleBody.getJsonArray(INSTANCES_KEY);
+
+    assertThat(allInstances.size(), is(2));
+    assertThat(sortedInstances.size(), is(2));
+    assertThat(retrieveBody.getInteger(TOTAL_RECORDS_KEY), is(2));
+
+    var firstInstance = allInstances.getJsonObject(0);
+    var secondInstance = allInstances.getJsonObject(1);
+    // no "sortBy" used so the database can return them in any order.
+    // swap if needed:
+    if (firstInstanceId.toString().equals(secondInstance.getString("id"))) {
+      var tmp = firstInstance;
+      firstInstance = secondInstance;
+      secondInstance = tmp;
+    }
+    final var sortedInstance = sortedInstances.getJsonObject(0);
+
+    assertThat(firstInstance.getString("id"), is(firstInstanceId.toString()));
+    assertThat(firstInstance.getString("title"), is("Long Way to a Small Angry Planet"));
+
+    assertThat(firstInstance.getJsonArray("identifiers").size(), is(1));
+    assertThat(firstInstance.getJsonArray("identifiers"),
+      hasItem(identifierMatches(UUID_ISBN.toString(), "9781473619777")));
+
+    assertThat(secondInstance.getString("id"), is(secondInstanceId.toString()));
+    assertThat(secondInstance.getString("title"), is("Nod"));
+
+    assertThat(secondInstance.getJsonArray("identifiers").size(), is(1));
+    assertThat(secondInstance.getJsonArray("identifiers"),
+      hasItem(identifierMatches(UUID_ASIN.toString(), "B01D1PLMDO")));
+
+    assertThat(sortedInstance.getString("title"), is("Long Way to a Small Angry Planet"));
+  }
+
+  @Test
   public void canSearchByClassificationNumberWithoutArrayModifier()
     throws InterruptedException, ExecutionException, TimeoutException {